RET-4718: Adding caseManagementLocation changes to caseTransfer
