+++ src/main/java/uk/gov/hmcts/ethos/replacement/docmosis/controllers/CaseActionsForCaseWorkerController.java
@@ -37,7 +37,7 @@ import uk.gov.hmcts.ethos.replacement.docmosis.service.CaseCloseValidator;
 import uk.gov.hmcts.ethos.replacement.docmosis.service.CaseCreationForCaseWorkerService;
 import uk.gov.hmcts.ethos.replacement.docmosis.service.CaseFlagsService;
 import uk.gov.hmcts.ethos.replacement.docmosis.service.CaseManagementForCaseWorkerService;
-import uk.gov.hmcts.ethos.replacement.docmosis.service.CaseManagementLocationCodeService;
+import uk.gov.hmcts.ethos.replacement.docmosis.service.CaseManagementLocationService;
 import uk.gov.hmcts.ethos.replacement.docmosis.service.CaseRetrievalForCaseWorkerService;
 import uk.gov.hmcts.ethos.replacement.docmosis.service.CaseUpdateForCaseWorkerService;
 import uk.gov.hmcts.ethos.replacement.docmosis.service.ClerkService;
@@ -108,7 +108,7 @@ public class CaseActionsForCaseWorkerController {
     private final NocRespondentRepresentativeService nocRespondentRepresentativeService;
     private final FeatureToggleService featureToggleService;
     private final CaseFlagsService caseFlagsService;
-    private final CaseManagementLocationCodeService caseManagementLocationCodeService;
+    private final CaseManagementLocationService caseManagementLocationService;
 
     private final NocRespondentHelper nocRespondentHelper;
 
@@ -295,7 +295,7 @@ public class CaseActionsForCaseWorkerController {
             log.info("HMC feature flag is {}", hmcToggle);
             if (hmcToggle) {
                 caseManagementForCaseWorkerService.setPublicCaseName(caseData);
-                caseManagementLocationCodeService.setCaseManagementLocationCode(caseData);
+                caseManagementLocationService.setCaseManagementLocationCode(caseData);
             }
         }
