+++ src/main/java/uk/gov/hmcts/ethos/replacement/docmosis/service/CaseManagementLocationService.java
@@ -1,16 +1,22 @@
 package uk.gov.hmcts.ethos.replacement.docmosis.service;
 
+import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
 import uk.gov.hmcts.ecm.common.model.helper.TribunalOffice;
 import uk.gov.hmcts.et.common.model.ccd.CaseData;
+import uk.gov.hmcts.et.common.model.ccd.types.CaseLocation;
+import uk.gov.hmcts.ethos.replacement.docmosis.domain.tribunaloffice.CourtLocations;
 
 import java.util.Objects;
+
+import static com.google.common.base.Strings.isNullOrEmpty;
 
 /**
  * This provides services to add caseManagementLocationCode to caseData.
  */
 @Service
-public class CaseManagementLocationCodeService {
+@Slf4j
+public class CaseManagementLocationService {
 
     private final TribunalOfficesService tribunalOfficesService;
 
@@ -18,7 +24,7 @@ public class CaseManagementLocationCodeService {
      * Standard constructor.
      * @param tribunalOfficesService used for location code lookups
      */
-    public CaseManagementLocationCodeService(TribunalOfficesService tribunalOfficesService) {
+    public CaseManagementLocationService(TribunalOfficesService tribunalOfficesService) {
         this.tribunalOfficesService = tribunalOfficesService;
     }
 
@@ -36,5 +42,23 @@ public class CaseManagementLocationCodeService {
         }
         String caseManagementLocationCode = tribunalOfficesService.getEpimmsIdLocationCode(tribunalOffice);
         caseData.setCaseManagementLocationCode(Objects.toString(caseManagementLocationCode, ""));
+    }
+
+    public void setCaseManagementLocation(CaseData caseData) {
+        String managingOfficeName = caseData.getManagingOffice();
+        if (isNullOrEmpty(managingOfficeName)) {
+            log.debug("leave `caseManagementLocation` blank since it may be the multiCourts case.");
+            return;
+        }
+
+        CourtLocations tribunalLocations = tribunalOfficesService.getTribunalLocations(managingOfficeName);
+        if (tribunalLocations.getEpimmsId().isBlank()) {
+            log.debug("leave `caseManagementLocation` blank since Managing office is un-assigned.");
+            return;
+        }
+        caseData.setCaseManagementLocation(CaseLocation.builder()
+                .baseLocation(tribunalLocations.getEpimmsId())
+                .region(tribunalLocations.getRegionId())
+                .build());
     }
 }
