+++ src/test/utils/java/uk/gov/hmcts/ethos/utils/ResourceUtils.java
@@ -1,0 +1,23 @@
+package uk.gov.hmcts.ethos.utils;
+
+import com.fasterxml.jackson.databind.ObjectMapper;
+import uk.gov.hmcts.et.common.model.ccd.CaseDetails;
+
+import java.io.IOException;
+import java.net.URISyntaxException;
+import java.nio.file.Files;
+import java.nio.file.Paths;
+import java.util.Objects;
+
+public final class ResourceUtils {
+    private ResourceUtils() {
+        // Access through static methods
+    }
+
+    public static CaseDetails generateCaseDetails(String jsonFileName) throws URISyntaxException, IOException {
+        String json = new String(Files.readAllBytes(Paths.get(Objects.requireNonNull(Thread.currentThread()
+                .getContextClassLoader().getResource(jsonFileName)).toURI())));
+        ObjectMapper mapper = new ObjectMapper();
+        return mapper.readValue(json, CaseDetails.class);
+    }
+}
